API: Move GateValve to top-level directory
